Add mock worker for local test/debugging rather than overloading the service metrics example with LLM specifics
