+++ applications/llm/count/Cargo.toml
@@ -37,3 +37,7 @@ opentelemetry = "0.20"
 opentelemetry-prometheus = "0.13"
 prometheus = "0.13"
 warp = "0.3"
+rand = "0.8"
+
+[dev-dependencies]
+reqwest = { version = "0.11", features = ["blocking"] }
